+++ tests/test_symmetric.py
@@ -5,8 +5,8 @@ from secrets import token_bytes
 
 import pytest
 
-from cwt.exceptions import DecodeError, VerifyError
-from cwt.key_types.symmetric import AESCCMKey, HMACKey, SymmetricKey
+from cwt.exceptions import DecodeError, EncodeError, VerifyError
+from cwt.key_types.symmetric import AESCCMKey, AESGCMKey, HMACKey, SymmetricKey
 
 
 class TestSymmetricKey:
@@ -47,10 +47,6 @@ class TestSymmetricKey:
             (
                 {1: []},
                 "kty(1) should be int or str(tstr).",
-            ),
-            (
-                {1: 4},
-                "k(-1) not found.",
             ),
             (
                 {1: 4, -1: 123},
@@ -103,6 +99,25 @@ class TestHMACKey:
         except Exception:
             pytest.fail("sign/verify should not fail.")
 
+    def test_hmac_key_constructor_with_hmac_256_256_without_key(self):
+        """"""
+        key = HMACKey(
+            {
+                1: 4,
+                3: 5,  # HMAC 256/256
+            }
+        )
+        assert key.kty == 4
+        assert key.kid is None
+        assert key.alg == 5
+        assert key.key_ops is None
+        assert key.base_iv is None
+        try:
+            sig = key.sign(b"Hello world!")
+            key.verify(b"Hello world!", sig)
+        except Exception:
+            pytest.fail("sign/verify should not fail.")
+
     @pytest.mark.parametrize(
         "invalid, msg",
         [
@@ -121,10 +136,6 @@ class TestHMACKey:
             (
                 {1: []},
                 "kty(1) should be int or str(tstr).",
-            ),
-            (
-                {1: 4},
-                "k(-1) not found.",
             ),
             (
                 {1: 4, -1: 123},
@@ -206,6 +217,56 @@ class TestAESCCMKey:
             pytest.fail("sign/verify should not fail.")
 
     @pytest.mark.parametrize(
+        "key_args, nonce",
+        [
+            (
+                {1: 4, 3: 10},
+                token_bytes(13),
+            ),
+            (
+                {1: 4, 3: 11},
+                token_bytes(13),
+            ),
+            (
+                {1: 4, 3: 12},
+                token_bytes(7),
+            ),
+            (
+                {1: 4, 3: 13},
+                token_bytes(7),
+            ),
+            (
+                {1: 4, 3: 30},
+                token_bytes(13),
+            ),
+            (
+                {1: 4, 3: 31},
+                token_bytes(13),
+            ),
+            (
+                {1: 4, 3: 32},
+                token_bytes(7),
+            ),
+            (
+                {1: 4, 3: 33},
+                token_bytes(7),
+            ),
+        ],
+    )
+    def test_aesccm_key_constructor_with_aes_ccm_without_key(self, key_args, nonce):
+        """"""
+        key = AESCCMKey(key_args)
+        assert key.kty == 4
+        assert key.kid is None
+        assert key.key_ops is None
+        assert key.base_iv is None
+        try:
+            encrypted = key.encrypt(b"Hello world!", nonce=nonce)
+            assert key.decrypt(encrypted, nonce) == b"Hello world!"
+        except Exception:
+            pytest.fail("sign/verify should not fail.")
+
+    @pytest.mark.parametrize(
         "invalid, msg",
         [
             (
@@ -223,10 +284,6 @@ class TestAESCCMKey:
             (
                 {1: []},
                 "kty(1) should be int or str(tstr).",
-            ),
-            (
-                {1: 4},
-                "k(-1) not found.",
             ),
             (
                 {1: 4, -1: 123},
@@ -293,6 +350,19 @@ class TestAESCCMKey:
             pytest.fail("AESCCMKey should fail.")
         assert msg in str(err.value)
 
+    def test_aesgcm_key_encrypt_without_msg(self):
+        key = AESCCMKey(
+            {
+                1: 4,
+                -1: token_bytes(16),
+                3: 10,  # AES-CCM-16-64-128
+            }
+        )
+        nonce = token_bytes(13)
+        with pytest.raises(EncodeError) as err:
+            key.encrypt(None, nonce=nonce)
+        assert "Failed to encrypt." in str(err.value)
+
     def test_aesccm_key_decrypt_with_invalid_nonce(self):
         """"""
         key = AESCCMKey(
@@ -332,3 +402,113 @@ class TestAESCCMKey:
         with pytest.raises(ValueError) as err:
             key.decrypt(encrypted, nonce=token_bytes(7))
         assert "The length of nonce should be 13 bytes." in str(err.value)
+
+
+class TestAESGCMKey:
+    """
+    Tests for AESGCMKey.
+    """
+
+    def test_aesgcm_key_constructor_with_aes_gcm_a128gcm(self):
+        """"""
+        key = AESGCMKey(
+            {
+                1: 4,
+                -1: token_bytes(16),
+                3: 1,  # A128GCM
+            }
+        )
+        assert key.kty == 4
+        assert key.kid is None
+        assert key.alg == 1
+        assert key.key_ops is None
+        assert key.base_iv is None
+        nonce = token_bytes(12)
+        try:
+            encrypted = key.encrypt(b"Hello world!", nonce=nonce)
+            assert key.decrypt(encrypted, nonce) == b"Hello world!"
+        except Exception:
+            pytest.fail("sign/verify should not fail.")
+
+    @pytest.mark.parametrize(
+        "key_args",
+        [
+            {1: 4, 3: 1},
+            {1: 4, 3: 2},
+            {1: 4, 3: 3},
+        ],
+    )
+    def test_aesgcm_key_constructor_with_aes_ccm_without_key(self, key_args):
+        """"""
+        key = AESGCMKey(key_args)
+        assert key.kty == 4
+        assert key.kid is None
+        assert key.key_ops is None
+        assert key.base_iv is None
+        nonce = token_bytes(12)
+        try:
+            encrypted = key.encrypt(b"Hello world!", nonce=nonce)
+            assert key.decrypt(encrypted, nonce) == b"Hello world!"
+        except Exception:
+            pytest.fail("sign/verify should not fail.")
+
+    @pytest.mark.parametrize(
+        "invalid, msg",
+        [
+            (
+                {1: 4, -1: b"mysecret", 3: 4},
+                "Unsupported or unknown alg(3) for AES GCM: 4",
+            ),
+            (
+                {1: 4, -1: b"mysecret", 3: 1},
+                "The length of A128GCM key should be 16 bytes.",
+            ),
+            (
+                {1: 4, -1: b"mysecret", 3: 2},
+                "The length of A192GCM key should be 24 bytes.",
+            ),
+            (
+                {1: 4, -1: b"mysecret", 3: 3},
+                "The length of A256GCM key should be 32 bytes.",
+            ),
+        ],
+    )
+    def test_aesgcm_key_constructor_with_invalid_args(self, invalid, msg):
+        """"""
+        with pytest.raises(ValueError) as err:
+            AESGCMKey(invalid)
+            pytest.fail("AESGCMKey should fail.")
+        assert msg in str(err.value)
+
+    def test_aesgcm_key_encrypt_with_empty_nonce(self):
+        """"""
+        key = AESGCMKey(
+            {
+                1: 4,
+                -1: token_bytes(16),
+                3: 1,  # A128GCM
+            }
+        )
+        with pytest.raises(EncodeError) as err:
+            key.encrypt(b"Hello world!", nonce=b"")
+        assert "Failed to encrypt." in str(err.value)
+
+    def test_aesgcm_key_decrypt_with_invalid_nonce(self):
+        """"""
+        key = AESGCMKey(
+            {
+                1: 4,
+                -1: token_bytes(16),
+                3: 1,  # A128GCM
+            }
+        )
+        assert key.kty == 4
+        assert key.kid is None
+        assert key.alg == 1
+        assert key.key_ops is None
+        assert key.base_iv is None
+        nonce = token_bytes(12)
+        encrypted = key.encrypt(b"Hello world!", nonce=nonce)
+        with pytest.raises(DecodeError) as err:
+            key.decrypt(encrypted, nonce=token_bytes(13))
+        assert "Failed to decrypt." in str(err.value)